+++ manta-crypto/src/arkworks/mod.rs
@@ -16,7 +16,6 @@
 
 //! Arkworks Backend
 
-pub use ark_ec as ec;
 pub use ark_r1cs_std as r1cs_std;
 pub use ark_relations as relations;
 pub use ark_snark as snark;
@@ -36,7 +35,9 @@ pub use ark_ed_on_bn254 as ed_on_bn254;
 
 pub mod algebra;
 pub mod constraint;
+pub mod ec;
 pub mod ff;
+pub mod glv;
 pub mod pairing;
 pub mod rand;
 pub mod ratio;