+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
+- [\#211](https://github.com/Manta-Network/manta-rs/pull/211) GLV scalar multiplication implementation
+- [\#286](https://github.com/Manta-Network/manta-rs/pull/286) MantaPay v1.0.0
 
 ### Changed